Aufteilung der Jelly-Komponenten und der Hauptklasse
