+++ src/main/java/jenkins/task/Exam.java
@@ -30,25 +30,14 @@
 package jenkins.task;
 
 import hudson.AbortException;
-import hudson.EnvVars;
 import hudson.Extension;
-import hudson.FilePath;
-import hudson.Launcher;
-import hudson.Launcher.ProcStarter;
 import hudson.Util;
 import hudson.model.AbstractProject;
-import hudson.model.Node;
-import hudson.model.Run;
-import hudson.model.TaskListener;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.Builder;
 import hudson.tools.ToolInstallation;
-import hudson.util.ArgumentListBuilder;
 import hudson.util.FormValidation;
 import hudson.util.ListBoxModel;
-import jenkins.internal.ClientRequest;
-import jenkins.internal.Remote;
-import jenkins.internal.data.FilterConfiguration;
 import jenkins.internal.data.ModelConfiguration;
 import jenkins.internal.data.ReportConfiguration;
 import jenkins.internal.data.TestConfiguration;
@@ -60,20 +49,12 @@ import jenkins.plugins.exam.config.ExamModelConfig;
 import jenkins.plugins.exam.config.ExamPluginConfig;
 import jenkins.plugins.exam.config.ExamReportConfig;
 import jenkins.plugins.shiningpanda.tools.PythonInstallation;
-import jenkins.task._exam.ExamConsoleAnnotator;
-import jenkins.task._exam.ExamConsoleErrorOut;
 import jenkins.task._exam.Messages;
-import jenkins.tasks.SimpleBuildStep;
-import org.apache.commons.lang.RandomStringUtils;
 import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 import org.kohsuke.stapler.QueryParameter;
 
-import javax.annotation.Nonnull;
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -82,45 +63,15 @@ import java.util.List;
  *
  * @author Kohsuke Kawaguchi
  */
-public class Exam extends Builder implements SimpleBuildStep {
+public class Exam extends ExamTask {
 
-    private String hash = "";
 
     private String useExecutionFile;
-    /**
-     * Identifies {@link ExamTool} to be used.
-     */
-    private final String examName;
-
-    /**
-     * Identifies {@link PythonInstallation} to be used.
-     */
-    private final String pythonName;
 
     /**
      * Identifies {@link jenkins.plugins.exam.config.ExamModelConfig} to be used.
      */
     private String examModel;
-
-    /**
-     * Identifies {@link jenkins.plugins.exam.config.ExamReportConfig} to be used.
-     */
-    private final String examReport;
-
-    private boolean pdfReport;
-    private String pdfReportTemplate;
-    private String pdfSelectFilter;
-    private boolean pdfMeasureImages;
-
-    /**
-     * Definiert den Report Prefix
-     */
-    private String reportPrefix;
-
-    /**
-     * JAVA_OPTS if not null.
-     */
-    private String javaOpts;
 
     /**
      * Definiert den Pfad zum ExecutionFile
@@ -131,20 +82,6 @@ public class Exam extends Builder implements SimpleBuildStep {
      * Definiert die ModelConfiguration
      */
     private String modelConfiguration;
-
-    /**
-     * Definiert die default SystemConfiguration
-     */
-    private String systemConfiguration;
-
-    private List<TestrunFilter> testrunFilter = new ArrayList<TestrunFilter>();
-
-    private boolean logging;
-    private String loglevelTestCtrl = getDescriptor().getDefaultLogLevel();
-    private String loglevelTestLogic = getDescriptor().getDefaultLogLevel();
-    private String loglevelLibCtrl = getDescriptor().getDefaultLogLevel();
-
-    private boolean clearWorkspace;
 
     private String pathExecutionFile;
     private String pathPCode;
@@ -183,108 +120,8 @@ public class Exam extends Builder implements SimpleBuildStep {
     }
 
     @DataBoundSetter
-    public void setSystemConfiguration(String systemConfiguration) {
-        this.systemConfiguration = systemConfiguration;
-    }
-
-    @DataBoundSetter
     public void setUseExecutionFile(String useExecutionFile) {
         this.useExecutionFile = useExecutionFile;
-    }
-
-    public String getReportPrefix() {
-        return reportPrefix;
-    }
-
-    public boolean getPdfReport() {
-        return pdfReport;
-    }
-
-    @DataBoundSetter
-    public void setPdfReport(boolean pdfReport) {
-        this.pdfReport = pdfReport;
-    }
-
-    public String getPdfReportTemplate() {
-        return pdfReportTemplate;
-    }
-
-    @DataBoundSetter
-    public void setPdfReportTemplate(String pdfReportTemplate) {
-        this.pdfReportTemplate = pdfReportTemplate;
-    }
-
-    public String getPdfSelectFilter() {
-        return pdfSelectFilter;
-    }
-
-    @DataBoundSetter
-    public void setPdfSelectFilter(String pdfSelectFilter) {
-        this.pdfSelectFilter = pdfSelectFilter;
-    }
-
-    public boolean getPdfMeasureImages() {
-        return pdfMeasureImages;
-    }
-
-    @DataBoundSetter
-    public void setPdfMeasureImages(boolean pdfMeasureImages) {
-        this.pdfMeasureImages = pdfMeasureImages;
-    }
-
-    @DataBoundSetter
-    public void setReportPrefix(String reportPrefix) {
-        this.reportPrefix = reportPrefix;
-    }
-
-    public boolean getLogging() {
-        return logging;
-    }
-
-    @DataBoundSetter
-    public void setLogging(boolean logging) {
-        this.logging = logging;
-    }
-
-    public List<TestrunFilter> getTestrunFilter() {
-        return testrunFilter;
-    }
-
-    @DataBoundSetter
-    public void setTestrunFilter(List<TestrunFilter> testrunFilter) {
-        this.testrunFilter = testrunFilter;
-    }
-
-    public String getLoglevelTestCtrl() {
-        return loglevelTestCtrl;
-    }
-
-    @DataBoundSetter
-    public void setLoglevelTestCtrl(String loglevelTestCtrl) {
-        this.loglevelTestCtrl = loglevelTestCtrl;
-    }
-
-    public String getLoglevelTestLogic() {
-        return loglevelTestLogic;
-    }
-
-    @DataBoundSetter
-    public void setLoglevelTestLogic(String loglevelTestLogic) {
-        this.loglevelTestLogic = loglevelTestLogic;
-    }
-
-    public String getLoglevelLibCtrl() {
-        return loglevelLibCtrl;
-    }
-
-    @DataBoundSetter
-    public void setLoglevelLibCtrl(String loglevelLibCtrl) {
-        this.loglevelLibCtrl = loglevelLibCtrl;
-    }
-
-    @DataBoundSetter
-    public void setClearWorkspace(boolean clearWorkspace) {
-        this.clearWorkspace = clearWorkspace;
     }
 
     @DataBoundSetter
@@ -302,237 +139,20 @@ public class Exam extends Builder implements SimpleBuildStep {
         this.systemConfiguration = Util.fixEmptyAndTrim(systemConfiguration);
     }
 
-    public String getExamName() {
-        return examName;
-    }
-
-    public String getPythonName() {
-        return pythonName;
-    }
-
     public String getExamModel() {
         return examModel;
-    }
-
-    public String getExamReport() {
-        return examReport;
     }
 
     public String getExecutionFile() {
         return executionFile;
     }
 
-    public String getSystemConfiguration() {
-        return systemConfiguration;
-    }
-
     public String getModelConfiguration() {
         return modelConfiguration;
     }
 
-    public boolean isClearWorkspace() {
-        return clearWorkspace;
-    }
-
-    /**
-     * Gets the EXAM to invoke, or null to invoke the default one.
-     */
-    public ExamTool getExam() {
-        for (ExamTool i : getDescriptor().getInstallations()) {
-            if (examName != null && examName.equals(i.getName())) {
-                return i;
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Gets the EXAM to invoke, or null to invoke the default one.
-     */
-    public PythonInstallation getPython() {
-        for (PythonInstallation i : getDescriptor().getPythonInstallations()) {
-            if (pythonName != null && pythonName.equals(i.getName())) {
-                return i;
-            }
-        }
-        return null;
-    }
-
     public ExamTool.DescriptorImpl getToolDescriptor() {
         return ToolInstallation.all().get(ExamTool.DescriptorImpl.class);
-    }
-
-    @DataBoundSetter
-    public void setJavaOpts(String javaOpts) {
-        this.javaOpts = Util.fixEmptyAndTrim(javaOpts);
-    }
-
-    /**
-     * Gets the JAVA_OPTS parameter, or null.
-     */
-    public String getJavaOpts() {
-        return javaOpts;
-    }
-
-    @Override
-    public void perform(@Nonnull Run<?, ?> run, @Nonnull FilePath workspace, @Nonnull Launcher launcher,
-                        @Nonnull TaskListener listener) throws InterruptedException, IOException {
-
-        ArgumentListBuilder args = new ArgumentListBuilder();
-
-        EnvVars env = run.getEnvironment(listener);
-
-        ExamTool examTool = getExam();
-        PythonInstallation python = getPython();
-        String exe = "EXAM.exe";
-        String pythonexe = "";
-        Node node = jenkins.internal.Util.workspaceToNode(workspace);
-        if (examTool == null) {
-            args.add("EXAM.exe");
-        } else {
-            if (node == null) {
-                throw new AbortException(Messages.EXAM_NodeOffline());
-            }
-            examTool = examTool.forNode(node, listener);
-            python = python.forNode(node, listener);
-            exe = examTool.getExecutable(launcher);
-            pythonexe = python.getHome();
-            if (!pythonexe.endsWith("exe")) {
-                if (!pythonexe.endsWith("\\") && !pythonexe.endsWith("/")) {
-                    pythonexe += File.separator;
-                }
-                pythonexe += "python.exe";
-            }
-            if (exe == null) {
-                throw new AbortException(Messages.EXAM_ExecutableNotFound(examTool.getName()));
-            }
-            args.add(exe);
-        }
-
-
-        File buildFile = new File(exe);
-        FilePath buildFilePath = new FilePath(buildFile);
-
-        String dataPath = examTool.getHome();
-        String configurationPath = null;
-        String examWorkspace = null;
-        String relativeDataPath = examTool.getRelativeConfigPath();
-        if (relativeDataPath != null && !relativeDataPath.trim().isEmpty()) {
-            dataPath = examTool.getHome() + File.separator + relativeDataPath;
-        }
-        configurationPath = dataPath + File.separator + "configuration";
-        examWorkspace = workspace + File.separator + "workspace_exam_restApi";
-        FilePath source = workspace.child("workspace_exam_restApi");
-        FilePath target = workspace.child("target");
-        examWorkspace = examWorkspace.replaceAll("[\\/]]", File.separator);
-        File configurationFile = new File(
-                dataPath + File.separator + "configuration" + File.separator + "config.ini");
-        if (!Remote.fileExists(launcher, configurationFile)) {
-            throw new AbortException(Messages.EXAM_NotExamConfigDirectory(configurationFile.getPath()));
-        }
-
-        if (workspace != null) {
-            args.add("-data", examWorkspace);
-        }
-        if (configurationPath != null) {
-            args.add("-configuration", configurationPath);
-        }
-
-        if (examTool != null) {
-            examTool.buildEnvVars(env);
-        }
-
-        ExamPluginConfig examPluginConfig = Jenkins.getInstance().getDescriptorByType(ExamPluginConfig.class);
-        int port = examPluginConfig.getPort();
-        args.add("--launcher.appendVmargs", "-vmargs", "-DUSE_CONSOLE=true", "-DRESTAPI=true",
-                "-DRESTAPI_PORT=" + port);
-
-        if(!examPluginConfig.getLicenseHost().isEmpty() && examPluginConfig.getLicensePort() != 0){
-            args.add("-DLICENSE_PORT=" + examPluginConfig.getLicensePort(), "-DLICENSE_HOST=" + examPluginConfig.getLicenseHost());
-        }
-
-        if (javaOpts != null) {
-            env.put("JAVA_OPTS", env.expand(javaOpts));
-            args.add(javaOpts.split(" "));
-        }
-
-        if (!launcher.isUnix()) {
-            args = toWindowsCommand(args.toWindowsCommand());
-        }
-
-        long startTime = System.currentTimeMillis();
-        try {
-            ExamConsoleAnnotator eca = new ExamConsoleAnnotator(listener.getLogger(), run.getCharset());
-            ExamConsoleErrorOut examErr = new ExamConsoleErrorOut(listener.getLogger());
-            boolean ret = true;
-            String slaveIp = Remote.getIP(launcher);
-            ClientRequest clientRequest = new ClientRequest(listener.getLogger(),
-                    "http://" + slaveIp + ":" + port + "/examRest");
-            try {
-
-                ProcStarter process = launcher.launch().cmds(args).envs(env).pwd(buildFilePath.getParent());
-                if (clientRequest.isApiAvailable()) {
-                    listener.getLogger().println("ERROR: EXAM is allready running");
-                    throw new AbortException("ERROR: EXAM is allready running");
-                }
-                process.stderr(examErr);
-                process.stdout(eca);
-                process.start();
-
-                ret = clientRequest.connectClient(5 * 60 * 1000);
-                if (ret) {
-                    TestConfiguration tc = createTestConfiguration();
-                    tc.setPythonPath(pythonexe);
-                    FilterConfiguration fc = new FilterConfiguration();
-
-                    for (TestrunFilter filter : testrunFilter) {
-                        fc.addTestrunFilter(new jenkins.internal.data.TestrunFilter(filter.name, filter.value,
-                                Boolean.valueOf(filter.adminCases), Boolean.valueOf(filter.activateTestcases)));
-                    }
-
-                    if (isClearWorkspace()) {
-                        clientRequest.clearWorkspace(tc.getModelProject().getModelName());
-                    }
-                    clientRequest.clearWorkspace(tc.getReportProject().getProjectName());
-                    if (!testrunFilter.isEmpty()) {
-                        clientRequest.setTestrunFilter(fc);
-                    }
-                    clientRequest.startTestrun(tc);
-
-                    clientRequest.waitForTestrunEnds(run.getExecutor());
-                    clientRequest.convert(tc.getReportProject().getProjectName());
-
-                    hash = "__" + RandomStringUtils.random(5, "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789".toCharArray());
-                    source = source.child("reports").child(tc.getReportProject().getProjectName()).child("junit");
-                    target = target.child("test-reports").child(tc.getModelProject().getProjectName() + hash);
-                    source.copyRecursiveTo(target);
-                }
-            } catch (Exception e) {
-                throw new AbortException("ERROR: " + e.getMessage());
-            } finally {
-                eca.forceEol();
-                clientRequest.disconnectClient(60 * 1000);
-            }
-            return;
-        } catch (IOException e) {
-            Util.displayIOException(e, listener);
-
-            String errorMessage = Messages.EXAM_ExecFailed();
-            if (examTool == null && (System.currentTimeMillis() - startTime) < 1000) {
-                if (getDescriptor().getInstallations() == null)
-                // looks like the user didn't configure any EXAM
-                // installation
-                {
-                    errorMessage += Messages.EXAM_GlobalConfigNeeded();
-                } else
-                // There are EXAM installations configured but the project
-                // didn't pick it
-                {
-                    errorMessage += Messages.EXAM_ProjectConfigNeeded();
-                }
-            }
-            throw new AbortException(errorMessage);
-        }
     }
 
     private ExamModelConfig getModel(String name) {
@@ -544,16 +164,7 @@ public class Exam extends Builder implements SimpleBuildStep {
         return null;
     }
 
-    private ExamReportConfig getReport(String name) {
-        for (ExamReportConfig rConfig : getDescriptor().getReportConfigs()) {
-            if (rConfig.getName().equalsIgnoreCase(name)) {
-                return rConfig;
-            }
-        }
-        return null;
-    }
-
-    private TestConfiguration createTestConfiguration() throws AbortException {
+    TestConfiguration addDataToTestConfiguration(TestConfiguration tc) throws AbortException {
         ModelConfiguration mod = new ModelConfiguration();
         ExamModelConfig m = getModel(examModel);
         if (m == null) {
@@ -564,69 +175,9 @@ public class Exam extends Builder implements SimpleBuildStep {
         mod.setTargetEndpoint(m.getTargetEndpoint());
         mod.setModelConfigUUID(modelConfiguration);
 
-        ReportConfiguration rep = new ReportConfiguration();
-        ExamReportConfig r = getReport(examReport);
-        rep.setProjectName(r.getName());
-        rep.setDbHost(r.getHost());
-        rep.setDbPassword(r.getDbPass());
-        rep.setDbPort(Integer.valueOf(r.getPort()));
-        rep.setDbSchema(r.getSchema());
-        rep.setDbService(r.getServiceOrSid());
-        rep.setDbType(r.getDbType());
-        rep.setDbUser(r.getDbUser());
-
-        TestConfiguration tc = new TestConfiguration();
         tc.setModelProject(mod);
-        tc.setReportProject(rep);
-        tc.setModelConfig("");
-        tc.setSystemConfig(systemConfiguration);
-        tc.setTestObject(executionFile);
-        tc.setReportPrefix(reportPrefix);
-
-        if (pdfReport && !pdfReportTemplate.isEmpty()) {
-            tc.setPdfReportTemplate(pdfReportTemplate);
-            tc.setPdfSelectFilter(pdfSelectFilter);
-            tc.setPdfMeasureImages(pdfMeasureImages);
-        }
-
-        tc.setLogLevelTC(RestAPILogLevelEnum.valueOf(loglevelTestCtrl));
-        tc.setLogLevelTL(RestAPILogLevelEnum.valueOf(loglevelTestLogic));
-        tc.setLogLevelLC(RestAPILogLevelEnum.valueOf(loglevelLibCtrl));
 
         return tc;
-    }
-
-    /**
-     * Backward compatibility by checking the number of parameters
-     */
-    protected static ArgumentListBuilder toWindowsCommand(ArgumentListBuilder args) {
-        List<String> arguments = args.toList();
-
-        // branch for core equals or greater than 1.654
-        boolean[] masks = args.toMaskArray();
-        // don't know why are missing single quotes.
-
-        args = new ArgumentListBuilder();
-        args.add(arguments.get(0), arguments.get(1)); // "cmd.exe", "/C",
-        // ...
-
-        int size = arguments.size();
-        for (int i = 2; i < size; i++) {
-            String arg = arguments.get(i).replaceAll("^(-D[^\" ]+)=$", "$0\"\"");
-
-            if (masks[i]) {
-                args.addMasked(arg);
-            } else {
-                args.add(arg);
-            }
-        }
-
-        return args;
-    }
-
-    @Override
-    public DescriptorImpl getDescriptor() {
-        return (DescriptorImpl) super.getDescriptor();
     }
 
     @Extension
@@ -680,15 +231,29 @@ public class Exam extends Builder implements SimpleBuildStep {
                     .getModelConfigs();
         }
 
+        private List<ExamReportConfig> addNoReport(List<ExamReportConfig> reports){
+            List<ExamReportConfig> lReportConfigs = reports;
+            boolean found = false;
+            for(ExamReportConfig config : reports){
+                if(config.getName().compareTo(ReportConfiguration.NO_REPORT) == 0){
+                    found = true;
+                    break;
+                }
+            }
+            if(!found){
+                ExamReportConfig noReport = new ExamReportConfig();
+                noReport.setName(ReportConfiguration.NO_REPORT);
+                noReport.setSchema("");
+                noReport.setHost("");
+                noReport.setPort("0");
+                lReportConfigs.add(0, noReport);
+            }
+            return lReportConfigs;
+        }
         public List<ExamReportConfig> getReportConfigs() {
             List<ExamReportConfig> lReportConfigs = Jenkins.getInstance().getDescriptorByType(ExamPluginConfig.class)
                     .getReportConfigs();
-            ExamReportConfig noReport = new ExamReportConfig();
-            noReport.setName(ReportConfiguration.NO_REPORT);
-            noReport.setSchema("");
-            noReport.setHost("");
-            noReport.setPort("0");
-            lReportConfigs.add(0, noReport);
+            lReportConfigs = addNoReport(lReportConfigs);
             return lReportConfigs;
         }
