--- src/main/resources/jenkins/task/Exam/help-campaignRuns_de.html
@@ -1,33 +1,0 @@
-<!--
-    Copyright (c) 2018 MicroNova AG
-    All rights reserved.
-
-    Redistribution and use in source and binary forms, with or without modification,
-    are permitted provided that the following conditions are met:
-
-        1. Redistributions of source code must retain the above copyright notice, this
-           list of conditions and the following disclaimer.
-
-        2. Redistributions in binary form must reproduce the above copyright notice, this
-           list of conditions and the following disclaimer in the documentation and/or
-           other materials provided with the distribution.
-
-        3. Neither the name of MicroNova AG nor the names of its
-           contributors may be used to endorse or promote products derived from
-           this software without specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-    ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-    WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-    DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
-    ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-    (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-    LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
-    ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
--->
-<div>
-    Durch setzen der Campaign Runs wird die geladene TestCampaign entsprechend oft
-    ausgeführt. Wird der Paramter nicht gesetzt wird eins als Default gesetzt.
-</div>
