+++ src/main/resources/jenkins/task/ExamSubmodules/Report/help-pdfMeasureImages.html
