+++ src/main/resources/jenkins/task/ExamSubmodules/Report/help-pdfMeasureImages_de.html
