+++ src/main/resources/jenkins/task/ExamSubmodules/Testrunner/config.jelly
@@ -1,0 +1,31 @@
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
+         xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+
+    <f:section title="${%Testrun}">
+        <f:entry title="Python Version" field="pythonName">
+            <f:select />
+        </f:entry>
+        <f:entry title="${%test object}" field="executionFile">
+            <f:textbox />
+        </f:entry>
+        <f:entry title="${%SystemConfiguration}" field="systemConfiguration">
+            <f:textbox />
+        </f:entry>
+        <f:optionalBlock title="${%configure Testrun Filters}" inline="true">
+            <f:entry>
+                <f:repeatableProperty field="testrunFilter" header="Testrun Filter" add="Add filter" />
+            </f:entry>
+        </f:optionalBlock>
+        <f:optionalBlock title="${%configure logging}" inline="true">
+            <f:entry title="${%TEST_CTRL}" field="loglevelTestCtrl">
+                <f:select default="${descriptor.getDefaultLogLevel()}"/>
+            </f:entry>
+            <f:entry title="${%TEST_LOGIC}" field="loglevelTestLogic">
+                <f:select default="${descriptor.getDefaultLogLevel()}"/>
+            </f:entry>
+            <f:entry title="${%LIB_CTRL}" field="loglevelLibCtrl">
+                <f:select default="${descriptor.getDefaultLogLevel()}"/>
+            </f:entry>
+        </f:optionalBlock>
+    </f:section>
+</j:jelly>
